% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_experts.R
\name{.validate_experts}
\alias{.validate_experts}
\title{Validate Experts List}
\usage{
.validate_experts(experts)
}
\arguments{
\item{experts}{List of experts.}
}
\value{
Validated experts list.
}
\description{
Internal validation function for Group SWARA.
}
\details{
Validates whether the experts object is a properly
structured list containing the required elements for
each expert.
}
\keyword{internal}
