% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merec_weights.R
\name{merec_weights}
\alias{merec_weights}
\title{MEREC Method for Objective Weights}
\usage{
merec_weights(data, benefit)
}
\arguments{
\item{data}{Decision matrix.}

\item{benefit}{Logical vector indicating criterion type:
TRUE = benefit criterion, FALSE = cost criterion.}
}
\value{
A list containing:
\item{method}{Method name.}
\item{normalized_data}{Normalized decision matrix.}
\item{S}{Overall performance vector of alternatives.}
\item{Sprime}{Performance matrix after removing each criterion.}
\item{E}{Removal effect vector for each criterion.}
\item{weights}{Normalized criterion weights.}
}
\description{
Method based on the Removal Effects of Criteria (MEREC)
for objective determination of criterion weights.
}
\references{
Keshavarz-Ghorabaee, M., Amiri, M., Zavadskas, E. K.,
Turskis, Z., & Antucheviciene, J. (2021).
Determination of objective weights using a new method based on
the removal effects of criteria (MEREC).
Symmetry, 13(4), 525.
}
