% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topsis.R
\name{topsis}
\alias{topsis}
\title{TOPSIS Method}
\usage{
topsis(data, weights, benefit)
}
\arguments{
\item{data}{Decision matrix.}

\item{weights}{Numeric vector of criterion weights.}

\item{benefit}{Logical vector indicating benefit criteria.}
}
\value{
A list containing:
\item{scores}{Closeness coefficients}
\item{ranking}{Ranking of alternatives}
\item{result}{Summary table}
}
\description{
Technique for Order Preference by Similarity to Ideal Solution (TOPSIS).
}
\references{
Hwang, C. L., & Yoon, K. (1981).
Multiple Attribute Decision Making:
Methods and Applications.
}
